Move useWpcomStore to own file
